# swarm, swarm/storage: lower constants for faster tests