This PR addresses https://github.com/ethereum/go-ethereum/pull/17854 and is reducing test constants and iterations for faster tests. On my machine I get:

```
ok      github.com/ethereum/go-ethereum/swarm   13.105s
ok      github.com/ethereum/go-ethereum/swarm/storage   10.474s
```

vs previously

```
ok      github.com/ethereum/go-ethereum/swarm   30.858s
ok      github.com/ethereum/go-ethereum/swarm/storage   38.774s
```