@nonsense 2018-10-08
Next package to review should be `cmd/swarm` which has especially slow tests.